Move all exported symbols to tool.go
This makes it easier to keep match.go in sync with the upstream, since
everything there is a direct copy from cmd/go code.

The tool.go file now contains all of the package API and documentation.
