+++ tool.go
@@ -2,7 +2,28 @@
 // as a convenience to developers who want to write tools with similar semantics.
 package gotool
 
-// export functions as here to make it easier to keep the implementations up to date with upstream.
+import "go/build"
+
+// Export functions here to make it easier to keep the implementations up to date with upstream.
+
+var DefaultContext = Context{
+	BuildContext: build.Default,
+}
+
+type Context struct {
+	BuildContext build.Context
+}
+
+// ImportPaths returns the import paths to use for the given arguments.
+//
+// The path "all" is expanded to all packages in $GOPATH and $GOROOT.
+// The path "std" is expanded to all packages in the Go standard library.
+// The string "..." is treated as a wildcard within a path.
+// Relative import paths are not converted to full import paths.
+// If args is empty, a single element "." is returned.
+func (c *Context) ImportPaths(args []string) []string {
+	return c.importPaths(args)
+}
 
 // ImportPaths returns the import paths to use for the given arguments using default context.
 //
@@ -12,5 +33,5 @@ package gotool
 // Relative import paths are not converted to full import paths.
 // If args is empty, a single element "." is returned.
 func ImportPaths(args []string) []string {
-	return DefaultContext.ImportPaths(args)
+	return DefaultContext.importPaths(args)
 }
